Support for "parent" type hint implemented
PHP's type hint system is a little bit strange. You can use the parent keyword
as a method type hint, which is totally fine for classes.

  class Foo extends Bar {
      public function foo(parent $x) {}
  }

But you can also use the parent keyword as a type hint within an interface
definition, even when this interface has more then one parent.

  interface Foo extends Bar, Baz {
      function x(parent $x);
  }

This is valid, but you cannot use the reflection extension to access the type
hint of parameter $x, because it fails with an error. But when you declare
a class that has parent class you can execute those methods.

  class x extends stdClass implements Foo {
      function x(parent $x) {}
  }

But when your class does not define a parent the PHP engine fails with an
error message: "Cannot access parent:: ..." :(